refactor: organize category route tests by scope
