+++ backend/tests/routes/categories/__init__.py
@@ -1,0 +1,1 @@
+"""Category route tests"""
